// src/app/models/DTO/userRequestUpdateDTO.ts
export class UserRequesUpdatetDTO{
    id: number = 0;
    documentType: string= '';
    documentNumber: string = '';
    name: string = '';
    lastName: string = '';
    email: string = '';
    password: string = '';
    phone: string = '';
    addresss: string= '';
    userType: string = '';
    state: number = 1;
}